feat: organize the store
this will make tackling the weekly refactor less painful
